# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Directory of Organizations API",
        "description": "API для работы со справочником организаций, зданий и видов деятельности",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/building": {
            "get": {
                "tags": [
                    "Buildings"
                ],
                "summary": "Получить список всех зданий",
                "operationId": "c505c1bd2088afcd7054d170a4f16450",
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Building"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Buildings"
                ],
                "summary": "Создать новое здание",
                "operationId": "b66f1a1addabf999897ee1afc7f8e85a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BuildingStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Здание создано",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Building"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        },
        "/api/building/{buildingId}": {
            "get": {
                "tags": [
                    "Buildings"
                ],
                "summary": "Получить здание по ID",
                "operationId": "9d5cac07bebb556ea29e18e599d00d65",
                "parameters": [
                    {
                        "name": "buildingId",
                        "in": "path",
                        "description": "ID здания",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Building"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Здание не найдено"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Buildings"
                ],
                "summary": "Удалить здание",
                "operationId": "c23ac9a86648d30b8d5c870cbb44134b",
                "parameters": [
                    {
                        "name": "buildingId",
                        "in": "path",
                        "description": "ID здания",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Здание удалено"
                    },
                    "404": {
                        "description": "Здание не найдено"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Buildings"
                ],
                "summary": "Обновить здание",
                "operationId": "e4a7be55dcd61e5ac23f200627c50028",
                "parameters": [
                    {
                        "name": "buildingId",
                        "in": "path",
                        "description": "ID здания",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BuildingUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Здание обновлено",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Building"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Здание не найдено"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Building": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "address": {
                        "type": "string",
                        "example": "г. Москва, ул. Тверская, 10"
                    },
                    "latitude": {
                        "type": "number",
                        "format": "float",
                        "example": 55.7558
                    },
                    "longitude": {
                        "type": "number",
                        "format": "float",
                        "example": 37.6176
                    }
                },
                "type": "object"
            },
            "BuildingStoreRequest": {
                "required": [
                    "address",
                    "latitude",
                    "longitude"
                ],
                "properties": {
                    "address": {
                        "type": "string",
                        "example": "г. Москва, ул. Тверская, 10"
                    },
                    "latitude": {
                        "type": "number",
                        "format": "float",
                        "example": 55.7558
                    },
                    "longitude": {
                        "type": "number",
                        "format": "float",
                        "example": 37.6176
                    }
                },
                "type": "object"
            },
            "BuildingUpdateRequest": {
                "properties": {
                    "address": {
                        "type": "string",
                        "example": "Обновленный адрес"
                    },
                    "latitude": {
                        "type": "number",
                        "format": "float",
                        "example": 55.7558
                    },
                    "longitude": {
                        "type": "number",
                        "format": "float",
                        "example": 37.6176
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "apiKey": {
                "type": "apiKey",
                "description": "Static API key authentication",
                "name": "x-api-key",
                "in": "header"
            }
        }
    },
...
}